KiraMeeT/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
asttokens==2.4.1
    # via
    #   -r requirements.in
    #   stack-data
build==1.2.2.post1
    # via
    #   -r requirements.in
    #   pip-tools
click==8.1.7
    # via
    #   -r requirements.in
    #   pip-tools
decorator==5.1.1
    # via
    #   -r requirements.in
    #   ipython
django==5.1.1
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-extensions
    #   django-rest-auth
    #   djangorestframework
    #   djangorestframework-simplejwt
django-allauth==65.0.2
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-rest-auth==0.9.5
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   django-rest-auth
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
exceptiongroup==1.2.2
    # via
    #   -r requirements.in
    #   ipython
executing==2.1.0
    # via
    #   -r requirements.in
    #   stack-data
ipython==8.28.0
    # via -r requirements.in
jedi==0.19.1
    # via
    #   -r requirements.in
    #   ipython
matplotlib-inline==0.1.7
    # via
    #   -r requirements.in
    #   ipython
packaging==24.1
    # via
    #   -r requirements.in
    #   build
parso==0.8.4
    # via
    #   -r requirements.in
    #   jedi
pexpect==4.9.0
    # via
    #   -r requirements.in
    #   ipython
pip-tools==7.4.1
    # via -r requirements.in
prompt-toolkit==3.0.48
    # via
    #   -r requirements.in
    #   ipython
psycopg2==2.9.9
    # via -r requirements.in
ptyprocess==0.7.0
    # via
    #   -r requirements.in
    #   pexpect
pure-eval==0.2.3
    # via
    #   -r requirements.in
    #   stack-data
pygments==2.18.0
    # via
    #   -r requirements.in
    #   ipython
pyjwt==2.9.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
pyproject-hooks==1.2.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
six==1.16.0
    # via
    #   -r requirements.in
    #   asttokens
    #   django-rest-auth
sqlparse==0.5.1
    # via
    #   -r requirements.in
    #   django
stack-data==0.6.3
    # via
    #   -r requirements.in
    #   ipython
tomli==2.0.2
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
traitlets==5.14.3
    # via
    #   -r requirements.in
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   asgiref
    #   ipython
wcwidth==0.2.13
    # via
    #   -r requirements.in
    #   prompt-toolkit
wheel==0.44.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
